Complete ISSUE #87
Create a new Component to set ZOOM Allowance
And if ZOOM burn fee is bigger than ZOOM allowance, it asks user to increase the ZOOM allowance.

Fixed some styles for mobile responsive.
